initial implementation of a DBus interface module
the most complicated thing is a ability to spawn a system process
and sent it stderr and stdout line-by-line to dbus as a simple event

for example start a module with `python3 modules/dbus_interface.py`
then call it method to start watchin for tmux session
```
dbus-send \
  --session \
  --type=method_call /manager \
  --dest=org.mindwm \
  mindwm.client.Run string:"tmux -Ltest -C a -t omg-tests" int32:1
```
where the first string parameter as a command to spawn and second int32
is a internal for mindwm stream id.

After that all tmux events on `test` socket of the `omg-test` session
will be published as a dbus events.