allow to define a regex to filter process output
This may be useful in case if we want to receive only system events
but not `%output` events. Such bevavior can be acheived via:

```
dbus-send --session \
  --type=method_call /manager \
  --dest=org.mindwm \
  mindwm.client.Run \
  string:"tmux -Ltest -C a -t omg-tests" \
  string:'^(?!%output.*)'
```

The last string parameter is a python regex which will be matched
against each line of process output and event will be sent only if
matched. To disable filter and catch all lines just provide an empty
string as a regex. I.e. `string:''` will pass all output.